Move entities to client folder
