--- src/client/yml/entities/form/index.ts
@@ -1,3 +1,0 @@
-export * from "./ui";
-
-export * from "./model";
